[wip] starting api/jobs of scheduling
