+++ src/beetle-srv/src/api/devices.rs
@@ -16,50 +16,6 @@ struct MessagePayload {
   message: String,
 }
 
-/// The api wrapper around convenience types for the underlying layout kinds.
-#[derive(Debug, Deserialize)]
-#[serde(rename_all = "snake_case", tag = "beetle:kind", content = "beetle:content")]
-enum QueuePayloadKind {
-  /// Controls the lights.
-  Lights(bool),
-
-  /// Renders text.
-  Message(String),
-
-  /// Will queue a QR code render for the device.
-  Link(String),
-
-  /// Attempts to rename the device.
-  Rename(String),
-
-  /// Attempts to queue a message that will force redisplay of registration.
-  Registration,
-
-  /// Will update the ownership record model to be public. This should be combined with the
-  /// `MakePrivate` variant, but expressing with a boolean on some `Toggle` variant always leads to
-  /// the confusion of what `true` means.
-  MakePublic,
-
-  /// Will update the ownership record model to be private.
-  MakePrivate,
-
-  /// Predefined.
-  Away,
-
-  /// Clears screen.
-  Clear,
-}
-
-/// The api used to add various layouts to a device queue.
-#[derive(Debug, Deserialize)]
-#[serde(rename_all = "snake_case")]
-struct QueuePayload {
-  /// The id of the device.
-  device_id: String,
-  /// The contents of the message.
-  kind: QueuePayloadKind,
-}
-
 /// The schema of our api to registration of a device.
 #[derive(Debug, Deserialize)]
 struct RegistrationPayload {
@@ -104,88 +60,6 @@ pub struct DeviceInfoPayload {
 /// Parses the payload from our message api. This should live in the request handler.
 async fn parse_message(request: &mut tide::Request<super::worker::Worker>) -> tide::Result<MessagePayload> {
   request.body_json::<MessagePayload>().await
-}
-
-/// Route: message
-///
-/// Sends a message to the device.
-pub async fn queue(mut request: tide::Request<super::worker::Worker>) -> tide::Result {
-  let queue_payload = request.body_json::<QueuePayload>().await.map_err(|error| {
-    log::warn!("bad device message payload - {error}");
-    tide::Error::from_str(422, "bad-request")
-  })?;
-  log::info!("queue payload request received - {queue_payload:?}");
-
-  let worker = request.state();
-
-  let user = worker
-    .request_authority(&request)
-    .await?
-    .ok_or_else(|| {
-      log::warn!("no user found");
-      tide::Error::from_str(404, "missing-user")
-    })
-    .map_err(|error| {
-      log::warn!("unable to determine request authority - {error}");
-      error
-    })?;
-
-  if worker.user_access(&user.oid, &queue_payload.device_id).await?.is_none() {
-    log::warn!("'{}' has no access to device '{}'", user.oid, queue_payload.device_id);
-    return Err(tide::Error::from_str(400, "not-found"));
-  }
-
-  log::info!("user {user:?} creating message for device - {:?}", queue_payload.kind);
-
-  let layout = match &queue_payload.kind {
-    kind @ QueuePayloadKind::MakePublic | kind @ QueuePayloadKind::MakePrivate => {
-      let privacy = match kind {
-        QueuePayloadKind::MakePublic => crate::registrar::ownership::PublicAvailabilityChange::ToPublic,
-        QueuePayloadKind::MakePrivate => crate::registrar::ownership::PublicAvailabilityChange::ToPrivate,
-        _ => return Ok(tide::Error::from_str(422, "bad transition").into()),
-      };
-
-      let job = crate::registrar::RegistrarJob::set_public_availability(queue_payload.device_id.clone(), privacy);
-      let id = worker.queue_job(job).await?;
-
-      return tide::Body::from_json(&RegistrationResponse { id })
-        .map(|body| tide::Response::builder(200).body(body).build());
-    }
-    QueuePayloadKind::Registration => {
-      let job = crate::registrar::RegistrarJob::registration_scannable(queue_payload.device_id.clone());
-      let id = worker.queue_job(job).await?;
-
-      return tide::Body::from_json(&RegistrationResponse { id })
-        .map(|body| tide::Response::builder(200).body(body).build());
-    }
-    QueuePayloadKind::Rename(new_name) => {
-      let job = crate::registrar::RegistrarJob::rename_device(queue_payload.device_id.clone(), new_name.clone());
-      let id = worker.queue_job(job).await?;
-
-      return tide::Body::from_json(&RegistrationResponse { id })
-        .map(|body| tide::Response::builder(200).body(body).build());
-    }
-    QueuePayloadKind::Away => crate::rendering::RenderVariant::message("Busy"),
-    QueuePayloadKind::Lights(true) => crate::rendering::RenderVariant::on(),
-    QueuePayloadKind::Lights(false) => crate::rendering::RenderVariant::off(),
-    QueuePayloadKind::Clear => crate::rendering::RenderVariant::message(""),
-    QueuePayloadKind::Link(m) => crate::rendering::RenderVariant::scannable(m.as_str()),
-    QueuePayloadKind::Message(m) => crate::rendering::RenderVariant::message(m.as_str()),
-  };
-
-  let request_id = worker
-    .queue_render(&queue_payload.device_id, &user.oid, layout)
-    .await
-    .map_err(|error| {
-      log::warn!(
-        "unable to queue render for device '{}' -> '{error}'",
-        queue_payload.device_id
-      );
-      error
-    })?;
-
-  tide::Body::from_json(&RegistrationResponse { id: request_id })
-    .map(|body| tide::Response::builder(200).body(body).build())
 }
 
 /// Route: message
@@ -318,7 +192,7 @@ pub async fn info(request: tide::Request<super::worker::Worker>) -> tide::Result
     sent_messages: device_diagnostic.sent_messages.unwrap_or_default(),
   };
 
-  log::debug!("user '{}' fetched device '{}'", user.oid, info.id);
+  log::trace!("user '{}' fetched device '{}'", user.oid, info.id);
   tide::Body::from_json(&info).map(|body| tide::Response::builder(200).body(body).build())
 }
