+++ src/beetle-ui/src/Alert.elm
@@ -1,7 +1,9 @@
 module Alert exposing (Alert(..), view)
 
+import Button
 import Html
 import Html.Attributes as A
+import Icon
 
 
 type Alert
@@ -9,11 +11,14 @@ type Alert
     | Happy String
 
 
-view : Alert -> Html.Html a
-view alert =
+view : Alert -> a -> Html.Html a
+view alert message =
     case alert of
         Warning content ->
-            Html.div [ A.class "alert-warning" ] [ Html.text content ]
+            Html.div [ A.class "alert-warning flex items-center w-full" ]
+                [ Html.div [ A.class "flex-1" ] [ Html.text content ]
+                , Html.div [ A.class "ml-auto" ] [ Button.view (Button.SecondaryIcon Icon.Cancel message) ]
+                ]
 
         Happy content ->
             Html.div [] []
